refactor(schema) split loading of plugins into a separate module
Split all schema-related, non-DB-related parts of loading plugins
into a separate module, `kong.db.schema.plugin_loader`.

This is the logic that needs to be shared between loading plugins
for initializing the Plugins DAO and loading plugins for building
the declarative configuration schema.
