+++ kong/db/dao/plugins.lua
@@ -1,9 +1,7 @@
 local constants = require "kong.constants"
-local typedefs = require "kong.db.schema.typedefs"
 local utils = require "kong.tools.utils"
-local Entity = require "kong.db.schema.entity"
 local DAO = require "kong.db.dao"
-local MetaSchema = require "kong.db.schema.metaschema"
+local plugin_loader = require "kong.db.schema.plugin_loader"
 
 
 local Plugins = {}
@@ -11,11 +9,7 @@ local Plugins = {}
 
 local fmt = string.format
 local null = ngx.null
-local type = type
-local next = next
 local pairs = pairs
-local ipairs = ipairs
-local insert = table.insert
 local tostring = tostring
 local ngx_log = ngx.log
 local ngx_WARN = ngx.WARN
@@ -143,156 +137,45 @@ function Plugins:check_db_against_config(plugin_set)
 end
 
 
---- Check if a string is a parseable URL.
--- @param v input string string
--- @return boolean indicating whether string is an URL.
-local function validate_url(v)
-  if v and type(v) == "string" then
-    local url = require("socket.url").parse(v)
-    if url and not url.path then
-      url.path = "/"
-    end
-    return not not (url and url.path and url.host and url.scheme)
+local function load_plugin_handler(plugin)
+  -- NOTE: no version _G.kong (nor PDK) in plugins main chunk
+
+  local plugin_handler = "kong.plugins." .. plugin .. ".handler"
+  local ok, handler = utils.load_module_if_exists(plugin_handler)
+  if not ok then
+    return nil, plugin .. " plugin is enabled but not installed;\n" .. handler
   end
+
+  return handler
 end
 
 
---- Read a plugin schema table in the old-DAO format and produce a
--- best-effort translation of it into a plugin subschema in the new-DAO format.
--- @param name a string with the schema name.
--- @param old_schema the old-format schema table.
--- @return a table with a new-format plugin subschema; or nil and a message.
-local function convert_legacy_schema(name, old_schema)
-  local new_schema = {
-    name = name,
-    fields = {
-      config = {
-        type = "record",
-        required = true,
-        fields = {}
-      }
-    },
-    entity_checks = old_schema.entity_checks,
-  }
-  for old_fname, old_fdata in pairs(old_schema.fields) do
-    local new_fdata = {}
-    local new_field = { [old_fname] = new_fdata }
-    local elements = {}
-    for k, v in pairs(old_fdata) do
+local function load_plugin_entity_strategy(schema, db)
+  local Strategy = require(fmt("kong.db.strategies.%s", db.strategy))
+  local strategy, err = Strategy.new(db.connector, schema, db.errors)
+  if not strategy then
+    return nil, err
+  end
+  db.strategies[schema.name] = strategy
 
-      if k == "type" then
-        if v == "url" then
-          new_fdata.type = "string"
-          new_fdata.custom_validator = validate_url
+  local dao, err = DAO.new(db, schema, strategy, db.errors)
+  if not dao then
+    return nil, err
+  end
+  db.daos[schema.name] = dao
+end
 
-        elseif v == "table" then
-          if old_fdata.schema and old_fdata.schema.flexible then
-            new_fdata.type = "map"
-          else
-            new_fdata.type = "record"
-            new_fdata.required = true
-          end
 
-        elseif v == "array" then
-          new_fdata.type = "array"
-          elements.type = "string"
-          -- FIXME stored as JSON in old db
-
-        elseif v == "timestamp" then
-          new_fdata = typedefs.timestamp
-
-        elseif v == "string" then
-          new_fdata.type = v
-          new_fdata.len_min = 0
-
-        elseif v == "number"
-            or v == "boolean" then
-          new_fdata.type = v
-
-        else
-          return nil, "unkown legacy field type: " .. v
-        end
-
-      elseif k == "schema" then
-        local rfields, err = convert_legacy_schema("fields", v)
-        if err then
-          return nil, err
-        end
-        rfields = rfields.fields.config.fields
-
-        if v.flexible then
-          new_fdata.keys = { type = "string" }
-          new_fdata.values = {
-            type = "record",
-            required = true,
-            fields = rfields,
-          }
-        else
-          new_fdata.fields = rfields
-          local rdefault = {}
-          local has_default = false
-          for _, field in ipairs(rfields) do
-            local fname = next(field)
-            local fdata = field[fname]
-            if fdata.default then
-              rdefault[fname] = fdata.default
-              has_default = true
-            end
-          end
-          if has_default then
-            new_fdata.default = rdefault
-          end
-        end
-
-      elseif k == "immutable" then
-        -- FIXME really ignore?
-        ngx_log(ngx_DEBUG, "Ignoring 'immutable' property")
-
-      elseif k == "enum" then
-        if old_fdata.type == "array" then
-          elements.one_of = v
-        else
-          new_fdata.one_of = v
-        end
-
-      elseif k == "default"
-          or k == "required"
-          or k == "unique" then
-        new_fdata[k] = v
-
-      elseif k == "func" then
-        -- FIXME some should become custom validators, some entity checks
-        new_fdata.custom_validator = nil -- v
-
-      elseif k == "new_type" then
-        new_field[old_fname] = v
-        break
-
-      else
-        return nil, "unknown legacy field attribute: " .. require"inspect"(k)
-      end
-
-    end
-    if new_fdata.type == "array" then
-      new_fdata.elements = elements
-    end
-    if new_fdata.type == nil then
-      new_fdata.type = "string"
+local function plugin_entity_loader(db)
+  return function(plugin, schema_def)
+    ngx_log(ngx_DEBUG, fmt("Loading custom plugin entity: '%s.%s'", plugin, schema_def.name))
+    local schema, err, err_t = plugin_loader.load_entity_schema(plugin, schema_def)
+    if not schema then
+      return nil, err, err_t
     end
 
-    insert(new_schema.fields.config.fields, new_field)
+    load_plugin_entity_strategy(schema, db)
   end
-
-  if old_schema.no_route then
-    insert(new_schema.fields, { route = typedefs.no_route })
-  end
-  if old_schema.no_service then
-    insert(new_schema.fields, { service = typedefs.no_service })
-  end
-  if old_schema.no_consumer then
-    insert(new_schema.fields, { consumer = typedefs.no_consumer })
-  end
-  return new_schema
 end
 
 
@@ -311,42 +194,18 @@ function Plugins:load_plugin_schemas(plugin_set)
       ngx_log(ngx_WARN, "plugin '", plugin, "' has been deprecated")
     end
 
-    -- NOTE: no version _G.kong (nor PDK) in plugins main chunk
-
-    local plugin_handler = "kong.plugins." .. plugin .. ".handler"
-    local ok, handler = utils.load_module_if_exists(plugin_handler)
-    if not ok then
-      return nil, plugin .. " plugin is enabled but not installed;\n" .. handler
+    local handler, err = load_plugin_handler(plugin)
+    if not handler then
+      return nil, err
     end
 
-    local schema
-    local plugin_schema = "kong.plugins." .. plugin .. ".schema"
-    ok, schema = utils.load_module_if_exists(plugin_schema)
-    if not ok then
-      return nil, "no configuration schema found for plugin: " .. plugin
+    local schema, err, err_t = plugin_loader.load_subschema(self.schema, plugin)
+    if err_t then
+      kong.log.warn("schema for plugin '", plugin, "' is invalid: ",
+                    tostring(self.errors:schema_violation(err_t)))
     end
-
-    local err
-
-    if schema.name then
-      local err_t
-      ok, err_t = MetaSchema.MetaSubSchema:validate(schema)
-      if not ok then
-        kong.log.warn("schema for plugin '", plugin, "' is invalid: ",
-                      tostring(self.errors:schema_violation(err_t)))
-      end
-
-    else
-      schema, err = convert_legacy_schema(plugin, schema)
-      if err then
-        return nil, "failed converting legacy schema for " ..
-                    plugin .. ": " .. err
-      end
-    end
-
-    ok, err = Entity.new_subschema(self.schema, plugin, schema)
-    if not ok then
-      return nil, "error initializing schema for plugin: " .. err
+    if err then
+      return nil, err
     end
 
     if schema.fields.consumer and schema.fields.consumer.eq == null then
@@ -367,37 +226,11 @@ function Plugins:load_plugin_schemas(plugin_set)
     }
 
     if db.strategy then -- skip during tests
-      local has_daos, daos_schemas = utils.load_module_if_exists("kong.plugins." .. plugin .. ".daos")
-      if has_daos then
-        local Strategy = require(fmt("kong.db.strategies.%s", db.strategy))
-        local iterator = daos_schemas[1] and ipairs or pairs
-        for name, schema_def in iterator(daos_schemas) do
-          if name ~= "tables" and schema_def.name then
-            ngx_log(ngx_DEBUG, fmt("Loading custom plugin entity: '%s.%s'", plugin, schema_def.name))
-            local ok, err_t = MetaSchema:validate(schema_def)
-            if not ok then
-              return nil, fmt("schema of custom plugin entity '%s.%s' is invalid: %s",
-                plugin, schema_def.name,
-                tostring(db.errors:schema_violation(err_t)))
-            end
-            local schema, err = Entity.new(schema_def)
-            if not schema then
-              return nil, fmt("schema of custom plugin entity '%s.%s' is invalid: %s", plugin, schema_def.name,
-                err)
-            end
-            local strategy, err = Strategy.new(db.connector, schema, db.errors)
-            if not strategy then
-              return nil, err
-            end
-            db.strategies[schema.name] = strategy
-
-            local dao, err = DAO.new(db, schema, strategy, db.errors)
-            if not dao then
-              return nil, err
-            end
-            db.daos[schema.name] = dao
-          end
-        end
+      local _, err, err_t = plugin_loader.load_entities(plugin, plugin_entity_loader(db))
+      if err_t then
+        return nil, fmt("%s: %s", err, tostring(db.errors:schema_violation(err_t)))
+      elseif err then
+        return nil, err
       end
     end
   end
